requirements.txt
diskcache==5.6.1
ffmpeg-python==0.2.0
future==0.18.3
Pillow==10.0.0
SQLAlchemy~=2.0.19